Move to grafana/data and add set options
