+++ packages/grafana-data/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './string';
+export * from './markdown';
